Remove govspeak dependency
This removes the dependency this gem has on govspeak. This is in
preparation of for govspeak to have the ability to call this gem as a
means to render components within govspeak. This avoids a cyclic dependency.

This gem previously used govspeak as a means to render markdown, there
wasn't actually any specific reason this was govspeak - probably
convenience more than anything. So I've stripped this back to use the
Kramdown gem directly.

I did debate whether to use commonmarker [1] instead of kramdown, but
felt that since there's a high chance govspeak will be
installed in places of this gem they could share the same markdown
library.

For styling I created a SASS mixin that could be used for markdown
typography and included in govspeak. This allows the styles to be
shared. I feel in time these may diverge slightly to allow features
govspeak disables like more headers, but this didn't seem necessary to
implement now.

For the wrapper for markdown HTML I went for a class of
`component-markdown` on the basis that it is intended to have the
results of markdown HTML inside it. I did consider using the term
"prose" instead of "markdown", as that is govuk-frontend terminology for
HTML without classes. However govuk-frontend have since removed this [2]
so I felt using the name "markdown" would be easier to understand.

As I put this together I actually think there's value to be unlocked
from having means to render markdown in apps and govspeak be two
separate things. We have to disable govspeak security features to
allow generic app rendering [3] so a divergence of these would likely
allow both govspeak security and in app HTML flexibility to increase.
Without the baggage of govspeak options we could also just allow
passing raw markdown direct to a component rather than having an app
convert it as is done with govspeak.

I've not taken these thoughts any further here though, I'll see how this
pans out and if these are real needs.

[1]: https://github.com/gjtorikian/commonmarker
[2]: https://github.com/alphagov/govuk-frontend/pull/778
[3]: https://github.com/alphagov/content-publisher/blob/654bf1a257e9a0de607e2c16b61754d55f5f8a72/app/helpers/application_helper.rb#L9