Permit Document List metadata fields to be blank
This will enable us to provide nil as a value in document list
metadata. E.g. public_updated_at: nil. This is useful, as
sometimes we can have nil fields and dont check for them,
which causes errors. We could fix this when using the component
but it is nicer to fix it once here.