Add gds-api-adapters and Plek as dependencies
GDS Api Adapters was a listed as a development dependency for this gem,
however it's required outside dev environment to pull in data from
Rummager [1]. Plek is also used here, so this is flagged as a
dependency.

Arguably both of these stretch beyond what this gem should actually be
doing, so shouldn't be encouraged patterns.

[1]: https://github.com/alphagov/govuk_publishing_components/blob/c1c602af10a3ee1dea24c77357c34b5afc843561/lib/govuk_publishing_components/presenters/services.rb