Allow spec_helper tests to run without loading Rails
This removes the rails loading from spec_helper and moves them only into
rails_helper. It then adds in the necessary requires so that the lib
tests of this app can run, this was tested with `rspec spec/lib`. To try
make this implicitly tested the spec_helper file is required before
Rails is included so hopefully any tests ran will show any introduced
require issues.

This is work to allow running this app outside of a Rails application,
namely with the govspeak gem. To do this we need to require the
appropriate files:

- active_support is required for mattr_accessor and various other
expectations of having Rails methods on objects
- rails/engine is required for loading the Rails::Engine class
- action_dispatch is also required for a Rails Engine.

Most likely govspeak will be used almost exclusively in Rails
applications, so the fact that an application needs to be bundled with
lots of Rails dependencies is of low concern (govspeak already bundles a
bunch of Rails dependencies such as actionview).