Add method to render components outside Rails
This adds a render method to GovukPublishingComponents which allows
rendering a component outside of a view context. This is added
specifically so that govspeak can call a component, but more generally
this allows calling components from anywhere in the code.

To allow this to work outside of a Rails environment the paths for views
and locale files are loaded in. In a Rails application these will
already be present.