Add inline_layout flag to button component
Without this flag, buttons next to each other in the code will visually display underneath each other because of a govuk-button style of width: 100% which is applied.

We want to use 2 buttons in the cookie banner, and need them to display next to each other so the banner takes up less space. The inline_layout flag sets the buttons to width: auto so they don't automatically take up 100% and display underneath each other.