Create attachment presenter
This presenter is intended to be a common place for rendering logic that
is shown as part of attachments. It ports logic that is currently already
in Whitehall [1] and Govspeak [2] with the intention that placing it here
may mean it won't have to be repeated again.

It is put together in conjunction with the Content Publisher work on
providing attachments [3]. It uses a content-type (MIME) approach to
file types rather than by extension and only supports types Content
Publisher will initially support.

The meta data available from this presenter is the data needed to
display an attachment link. These use ActionView helpers, hence the need
to require action_view in this gem.

The content types are stored in a hash which can have keys of
content_type, name, abbr, and extension. content_type and name are
required for a supported content type - the name being the value that
described the type of file. An abbr value can be specified to indicate
that a file type can be shown as an abbreviation. The extension value is
only necessary when files share a content type but behaving differently
depending on their extension. The extension value is used to determine
which file.

[1]: https://github.com/alphagov/whitehall/blob/5c760eea912320c5a80ef59791df47161d889f2f/app/helpers/document_helper.rb
[2]: https://github.com/alphagov/govspeak/blob/c3a0742c87537a371108d097cea23688efa90a14/lib/govspeak/presenters/attachment_presenter.rb
[3]: https://github.com/alphagov/content-publisher/commit/5ac7ecf19bcba675f98ddb6375566ae89000c2d3