Attachment link component
This ports the attachment link styling that is currently in govspeak and
Whitehall into govuk_publishing_components. It is based on [1].

This code is unlikely to be imminently removed from either Govspeak or
Whitehall as Govspeak uses it for Specialist Publisher attachments and
Whitehall uses it for it's own rendering. However it is hoped that in
time it can be eventually removed from both of these.

This component is somewhat unusual in that it's expected to inherit
styling from it's container. I felt this was better than specifying font
styling for it as this would cause problems if it was ever used in
sections with larger or smaller text. It ports over the same abbr styles
that are used in govspeak for consistency, as there is so few of them I
thought it was simpler to move them over than to create a mixin.

This component is somewhat intertwined with the work the Publishing
Workflow team are doing on attachments so I've marked this as
experimental while this is under development. The intention is to remove
the experimental flag once Content Publisher supports attachments for news
(hopefully by June 2019).

[1]: https://github.com/alphagov/govspeak/blob/c3a0742c87537a371108d097cea23688efa90a14/lib/govspeak.rb#L262-L269