Add Attachment component
This ports much of the styling from the Attachment meta component in Content
Publisher[1] with the intention that this component will largely replace the
attachment display part of it.

This component is intended to be used as a 'preview' of attachment markdown in
Content Publisher, which is why we display OpenDocument help text when
appropriate to represent the text that would appear on the frontend of GOV.UK.
However since we also intend to use it in the attachment meta component, which
will essentially display a list of attachments that will appear for users to
select to insert in the body of their content, we need to have the option not
to display this help text so that the list doesn't become cluttered or
confusing to publishers.

Since the thumbnail links to the same place as the text link, we add
`tabindex=1` to the thumbnail link so that users don't need to tab through two
identical links.

[1]:
https://github.com/alphagov/content-publisher/blob/master/app/assets/stylesheets/components/_attachment-meta.scss