Update cookies banner behaviour without Javascript
The banner was supposed to not show without Javascript. It did not
display in the component guide, but showed up on the frontend. However
we want the banner to still display, as there are other cookies that
may still be set by servers, and not Javascript.

We also don't show the "Hide this message" link without Javascript.