Add MVP consent cookie
We need to set a cookie that stores a user's preference for cookies.
This commit also introduces a setConsentCookie method that is not used yet, but will be used when we develop the cookie settings page.
The cookie settings page will let users allow/deny specific cookie categories, as opposed to just being able to allow or deny all types.