Add document? and spreadsheet? methods to attachment presenter
To allow determining which image to show for an attachment this adds
boolean methods that allow asking whether the attachment file type
belongs to a particular group. These groups are documents and
spreadsheets.

The groupings are based on logic from Whitehall: https://github.com/alphagov/whitehall/blob/5c760eea912320c5a80ef59791df47161d889f2f/app/helpers/document_helper.rb#L136-L148

However a difference is that PDF files will have a document image, this
is because a decision has been made to not allow PDFs to have a
thumbnail preview image anymore.

I felt it was simpler to do this by adding a flag to the list of types
rather than a secondary list of mime types. I then changed opendocument
to be done in the same way.

I removed the thumbnail method as this won't be used anymore.