Embed SVG icons in attachment component
This replaces the temporary image element with a choice of three SVG
files that are used to show a thumbnail based on the file content type.

These SVG's are embedded not only because they are lightweight, but also
to solve a tricky problem with govspeak. As govspeak can render this
component it means that the app that renders this HTML may not be the
same one that displays it (for example, Publishing API may render the govspeak yet
Government Frontend may display it). This means it is not safe to use
assets from asset pipeline as the ones Publishing API uses may not be
the same as Government Frontend has and, similarly, as this HTML may
live in the Content Store for a long time an asset fingerprinting is
also not safe. These issues are resolved by embedding the images
directly into the HTML.

These are SVG reproductions of the images used by Whitehall for
attachments, with the exception of the HTML one - as this component does
not support HTML attachments yet.

I've also updated the examples to include examples of each thumbnail.