Drop the readable_* methods on Attachment presenter
These weren't the best named methods, as it wasn't clear who they will
readable by and nor clear how a number_of_pages could be unreadable.

With a little bit of reflection I felt that these didn't need to be part
of the presenter, the view is already doing a good chunk of their
logic and it'd be more idiomatic to use view helpers in views.