Revert "Merge pull request #849 from alphagov/prevent-double-click-on-buttons"
This reverts commit 19cc9d09a152e4e38153c6d792e582a5f6c08b38, reversing
changes made to 9f259ca57bf08f141e30f035b4403d3331115c3a.

Previously we changed the button component to automatically prevent
double click for submit-type buttons. This caused an unexpected testing
issue in a consuming app (Content Publisher) in the context of a modal,
where multiple buttons are clicked in succession without a page reload.

This reverts the change to automatically prevent double click, pending
an iteration to the logic in govuk-frontend to scope the prevention to
the same button.