Add SearchAction for manuals
Manuals like [The Highway Code](https://www.gov.uk/guidance/the-highway-code)
have an embedded search box which allows searches scoped to that manual.

By advertising this in our machine readable metadata it's more obvious to
external consumers of the site that these things are available.

This is part of the work to make our search more visible to external consumers
of the site.

https://trello.com/c/4EAytHVV/747-plan-indexing-of-finder-pages