Check the consent cookie before setting cookies
We need to check what cookies a user has consented to before we set each cookie.

This also tidies up the cookie banner code so the old banner and new banner are more separate. This will make it easier for us to
remove the old banner code once the new banner has been deployed.