Remove step by steps from CreativeWork schema
Our structured data report indicates that these were not including mandatory
properties of the HowTo schema (that is, `name` and `step`). To include `step`
would mean including the whole of the step by step in each related CreativeWork's
schema which I'm not sure would be helpful.

In any case they seem to be actively hindering external crawlers, so let's
remove them for now.

We've fixed the [HowTo schema in Collections](https://github.com/alphagov/collections/pull/1124)
so hopefully the combination of these two things will have a positive result.