Add machine readable breadcrumbs for step by steps
We've removed the `partOf` metadata for step by steps which linked back
to the "parent" because we were using a partial HowTo schema which is
not supported by Google's search indexing process.

This adds in machine readable breadcrumbs for pages that show the step by
step header which should provide a better experience.