Add the closed step by step for secondary content
Make it possible for the step by step side bar to appear in the side navigation for all content that uses the contextual navigation.

In a separate card we will: https://trello.com/c/iE60uYXs/90-create-interface-in-collections-publisher-for-secondary-content

(Note: we will keep all content the same for the breadcrumbs and side nav as for step by steps elsewhere, we are not changing 'part of' for example.)

Update developer documentation with these changes.

**Logic:**

If the secondary content applies to more than 1 step by step then:
- list out links to step by steps in the side bar (use step by step sidebar styling)
- step by step 'breadcrumb' does not now

If the content is tagged to the business finder and is secondary content on a step by step then the secondary content takes precedence.

We want to add a new feature to 'step by step publisher' that makes it possible to add step by step navigation to secondary content. (Secondary content = content that is about the service but is not useful enough to include in the step by step).

**MVP = add the closed step by step to the side nav for these pages**

**Nice to have =  make is possible to assign a step in the step by step to open for that piece of secondary content, where relevant.**

Ref: https://trello.com/c/eC8EXI21/120-step-by-step-add-the-closed-step-by-step-to-the-side-nav-for-secondary-content