Use "Component Guide" as title instead of component_guide_title
We're currenlty using the component_guide_title defined in config for both the header, heading one and breadcrumb root. While the link in the header points to the root of the application, the breadcrumb root is pointing to the component guide route. This PR updates the template to keep the custom defined title only for the header and use a "Component Guide" for the rest.