Accept items on error_message component
There is a common pattern for collections of errors, known as
error_items. This an array of hashes where each entry may have a text
attribute of the error, this pattern is common as it is the same input
that is used for error summary.

The conversion of this pattern into input for error message is common
across this project with implementations of it in input component, date
component, textarea component and radio component. Apps, such as Content
Publisher, are even duplicating the code with in app code.

This level of repetition isn't ideal - especially as it deals with raw
HTML. So I've moved this to the error_message component to consolidate
this logic.

This also resolves a potential XSS issue as any item[:text] values will
be escaped unless they are previously marked with html_safe.