Add unsafe-eval to CSP output for component requests
The Axe-core library that govuk_publishing_components uses makes use of
a dependency that uses the eval function so this directive is needed
until that is removed.

Relevant project issues:

- https://github.com/dequelabs/axe-core/issues/1175
- https://github.com/olado/doT/issues/276

This effects of this issue is that all use of component guides trigger
CSP errors for every page view. Oddly this issue doesn't seem to be
observable in Chrome when report only is on, but can be seen in Firefox.