Fix alignment of text and buttons in cookie banner
Also simplifies the inline button CSS as it was getting overcomplicated. Makes the inline button styles more generic, so the button can be styled specifically for the cookie banner in the cookie banner css

govuk-buttons has a style of vertical-align: top which was making it difficult to align the buttons and the text in the cookie banner. Overriding this with vertical-align: baseline fixes this problem