Change cookie banner text to green
We want the cookie banner text to be green to match the buttons.

However, we don't want the confirmation text to be green as at this stage the banner has no buttons and we think all green text might look strange. For this reason, we're also keeping the banner border black.