Add ministers with roles to org schema
This adds the ministers of an organisation to the GovernmentOrganization
schema entry.  It includes their name, prefix (Rt Hon), url of people page
and image.

For each role they inhabit we also include the role name, and the url of
the role page.

I'm reimplementing the Person schema over reusing `person_schema.rb` here
because that is closely coupled to describing itself (that is, the people
page on which it sits.).

The ministers descibed on org pages are described in a different json format
to people pages and also have role information embedded.  There may be an
opportunity to do some refactoring when we get to implementing a similar
approach for speeches and other things also tagged to people, but I think
it's likely that org pages will remain "different" because of the structure
of their content items.

The content item for an organisation may include an `ordered_ministers`
array in the details hash.  (See the [ministry-of-defence](https://www.staging.publishing.service.gov.uk/api/content/government/organisations/ministry-of-defence)
for an example).

There is one entry per role, not per person, so we need to be able to
collapse the roles to avoid having more than one Theresa May in the Cabinet
Office schema for example.  (She's currently [the Prime Minister, the First
Lord of the Treasury and Minister for the Civil Service](https://www.gov.uk/government/organisations/cabinet-office))

We could do this in one `reduce`, but I think that might be more difficult
to understand than generating the people and roles in one step, and then
joining the schemas in a second step.

This passes the [structured data testing tool](https://search.google.com/structured-data/testing-tool/u/0/)
when tested.

https://trello.com/c/iNXeAeJc/146-to-add-the-person-schema-to-org-pages-on-govuk