Update feedback component error handling
At the moment, the alphagov/feedback application responds
with a 422 HTTP status if all fields are left empty in the form.

The feedback components javascript was setup to always apply the
constructed error message first and then check for the
`responseJSON.message` property, which would always return undefined.

This resulted in 422 status error message to never being applied.

This PR updates the logic to account for that.