Add new getConsentCookie method and JSON.parse fallback
When we want to check the value of the consent cookie, we have to parse the value returned by window.GOVUK.cookie('cookie_policy'). There was no function which did all of this, so we had to call JSON.parse in our other repos where necessary.

Our Jasmine tests do not parse the cookie value properly, so we were having to stub JSON.parse. This was ok to do as a workaround in one repo, but becomes difficult to manage when it has to be done in tests across multiple repos every time we want to check the consent cookie.

This commit adds a new method which fetches the consent cookie and parses it (sometimes twice if needed) so all the logic is kept within the cookie-functions helper in govuk_publishing_components.