Remove new cookie banner references
Some places in the code still reference the class 'gem-c-cookie-banner--new' because we had to deploy with that code in place due to caching and the way static and frontend apps interact.

It has been a week since we removed the old code, so `gem-c-cookie-banner--new` should no longer be being used anywhere.