Reset focus when user closes feedback form before submission
Currently if the user closes the "not useful" or "something is wrong" feedback
form the focus is lost. This is an accessibility issue as keyboard users end up
back at the start of the page, and screenreader users are unaware of where they
are on the page. This resets the focus back to the link that was activated to
open the form when it is closed.