Move tab style overrides to the gem
The tab styles are affected by the .content-block styles. In Frontend, there are overrides for this problem. However, they don't work properly with our current tabs on mobile (the left spacing is incorrect). This corrects the overrides and moves them into the gem, so they're in one common place and more visible to those making changes to the tabs component.