Fix duplicate link bug with feedback form
The hidden input for GA ID and the "Don't have an email address?" link are inserted using Javascript. This is done when the user clicks 'No' within the feedback form and the form opens. However, there is no check for whether the links have already been inserted. This means that the links are appended each time the user opens the 'No' form.

This commit adds a check so that we don't insert the links if they're already present