Only setup asset precompilation when assets are available
This resolves an issue where this gem is included into a Rails
application which doesn't have sprockets installed and then errors
because Rails.application.config.assets is undefined.

This is the case with Publishing API which uses govspeak to render HTML
(and thus implicitly depends on this gem) but does not have sprockets
included.