Hide test components in the dummy app
The dummy app in `spec/dummy` has 2 purposes.

First, it's a sample app that is used in tests to check the
functionality of having components inside the application.

Secondly, it is hosted on https://components.publishing.service.gov.uk
as  a reference.

This adds a environment variable to the startup scripts for the
reference app, so that we can hide the components that are only used
for testing.