Add FAQPage schema
What
----
This will be used on Guides to present the chapters in an improved way in
featured snippets.

Implemented as per: https://developers.google.com/search/docs/data-types/faqpage

Why
---
SEO improvement. It allows users to read more of the content in Google,
hopefully allowing them to make better decisions about whether to read more of
the content and thus shortening their journeys.

We will measure the effects of this change.