Fix FAQPage schemas with empty answers
Some pages have adjacent h2 tags which means that we will be generating
question and answer pairs that don't actually have an answer.

(An answer in an FAQPage schema is the html that resides after an h2 tag,
the "question" is the text of the h2 tag)

An example page is https://www.gov.uk/check-a-university-is-officially-recognised/listed-bodies
where there is an A-Z list with no content between the V and W h2 titles.