Use the page title as the default FAQPage question
Previously we used "Summary" to name a section of html that didn't have
an h2 tag preceding it.  Unfortunately, Google appears to not include
sections in rich results unless all the text actually appears on the page.

We still get rich snippets, we're just missing the first section.  Hopefully
this will be remedied by using the page title text instead.

This wasn't flagged by the rich text testing tool.