Add content security policy
This adds the GOV.UK security policy to the dummy application of this
gem. This means that when using this app in dev or viewing it at
https://components.publishing.service.gov.uk/component-guide the GOV.UK
CSP will apply.

This also has to unfortunately allow unsafe-eval for JS so that the
component-guide browser can use AXE. Relevant project issues:

- dequelabs/axe-core#1175
- olado/doT#276