Replace regex with similar method from Rails
https://apidock.com/rails/ActiveSupport/Inflector/parameterize

We want to be able to generate links to steps from other places on the site.
Using the rails builtin methods for doing this seems like an easy way to
converge - even if one of the existing tests does end up looking a little weird.

(I'm not sure that 'Step 3?"`_!' is a piece of text that we're likely to see
 for real.)