Refactor contextual guidance to refer the input
Refactor contextual guidance to specify that a guidance container belongs to an input field (should be displayed when that input is focused), rather than using data-attributes on inputs to specify which guidance should be shown on focus. This requires an `input_id` attribute to be passed to the component and removes the data attributes set on input fields.