Make it possible to render a document list without links
This is needed to support using the document list for roles in
Whitehall. Some roles don't have their own URLs (they are base path less
editions) so it doesn't make sense to render a link.

There is some confusion over whether the name of the component "document
list" makes sense with this feature. We've decided that it should be
fine as long as we make it clear that this is a list of documents, not a
list of links (as it is currently written in the description).