Allow aria-controls attribute
This is adding the ability to specify the aria-controls attribute,
that's available on the input component.

Why
On finders we are looking to upgrade the text input with a search icon
to a search component.

The latter is currently missing this attribute which is being used
to announce dynamic changes.