Change tags and styling for feedback component
This change addresses the issue encountered when a user that has blocked
css styling sees the feedback component. Originally, the options
presented to the user are all strung together in one line, making it
potentially confusing. This change separates them out into a list,
meaning that when there is no styling for the page, each option is on
its own separate row.

The styling is effectively neutralised if css is turned on, and all of
the default behaviours for the lists are removed.