Remove Brexit CTA link from checker start page
Currently we want to show a contextual Brexit sidebar on start pages for any content that is tagged to /brexit or /world/brexit.
This link serves as a call to action to bring people back to the Brexit landing page.
The landing page in turn links to the checker as a major feature, which itself has a start page.
Currently that start page shows this sidebar, putting the user in a loop between the start page and landing page.

To prevent this we add an additional condition to the show_brexit_cta? method, only used in this repo in _show_contextual_sidebar.html.erb,
causing it to return false if the content_id matches the content_id for the Brexit checker start page