Remove hard limit on step by steps
We initially had a hard limit on step by step pages to prevent pages that are
part of many routes from being overwhelmed with sidebar navigation. We envisaged
that https://www.gov.uk/request-copy-criminal-record might be an example of
this.

This was before the publishing tool / pattern were iterated meaning that
publishers had a mucher finer-grained control over when step by step navigation
was shown.

It appears that this limit may have run its course, so we're going to try
removing it for the time being.

https://govuk.zendesk.com/agent/tickets/3884434