Make step by steps resemble related link styles
Feedback from the govuk-design channel was that we need to increase the spacing
between the links to related step by step pages so that they were "like
related links".

This change also removes the bold font weight when multiple related step
by steps are shown in the sidebar.  It does not affect the font weight of
the title of a step by step if it is the only one (and therefore also renders
the full step by step outline).

Using normal font weight means that the links are closer to related links
in appearance, thus increasing the consistency of navigation styles.