Add special case for 'UK mission in the European Union'
The slug for the world location 'UK mission in the European Union'
doesn't match exactly the title of the world location. Instead the slug
shortens 'European Union' to 'EU'.

This means that links to this world location are currently broken (for
example
https://www.gov.uk/government/speeches/triggering-the-jcpoa-dispute-resolution-mechanism-foreign-secretarys-commons-statement
(see section at the bottom of the page).

This is similar to the special cases that we used to have in
government-frontend before they were removed and moved into
govuk_publishing_components:
https://github.com/alphagov/government-frontend/pull/1035