Add alternative header layout
Add a new layout alternative that allows for search on the left
of the header and navigation items on the right to support some
layouts seen in whitehall pages.
Correct a stray div that was breaking the with-nav option
without a searchbox.
Add missing govuk-grid-row wrapper for govuk grid columns.
Add tracking to home link.