Add placeholder text if label is inline
By default the label in inline, unless the `inline_label`
property is set to false.

At the moment if the input has a value and and inline label,
the two overlap until the input receives focus.

This change introduces a visually hidden label when it is set to be inline
and uses the placeholder property to display the label text.

Placeholder itself is not accessible/read by screenreader,
so we use the visually hidden label to annouce the input's intent.

If inline_label property is set to false, we display the label above
the input and placeholder text is not rendered