Remove font-size option for contents list
We added this option to the component for one iteration of the Brexit landing page, but it ended up not being used.

The mobile team are looking at making font-sizes more consistent across GOVUK so page hierarchy is clearer. The font-size flag on this component isn't being used yet and we probably don't want to encourage it's use. We may as well remove it for now.