Allow collapsing on mobile to be switched off
In #1290, we changed the default behaviour of the breadcrumb component to collapse down to the first and last item on mobile. This is part of a change to make breadcrumbs consistent across GOV.UK. However, there are some cases of breadcrumb use at the moment which would break with this new behaviour.

This commit adds a new flag which allows us to override this default collapsing behaviour if needed.