Fix tests and general refactor
This branch inherited new component functionality after a rebase.
Since the code has diverged from master the expected results of
the tests of the new functionality differ, so this updates them
to pass.

Also fixes some linting, indentation and spacing issues and removes a
typo.

Refactors the way classes are applied on the layout header and search
components.

Updates layout_header.scss to request the styles for search and
skip-link explicitly. This is necessary due to a change in the way CSS
is called.