Initialise global bar on approving all cookies
Currently, if a user accepts cookies and then immediately clicks to hide the global banner, this setting isn't saved and they will see the banner the next time they refresh.
This is because the global_bar_seen cookie isn't set when giving consent to cookies, it is only set on the next page view.

This commit runs the initialise code for the global bar when cookies are approved, to set the cookie without the need for page refresh.