Add 'hide_order_copy_link' parameter to hide 'Order a copy'
Attachment is used on admin-facing pages as well as user-facing
pages. Their requirements for a preview are somewhat different,
and having 'Order a copy' doesn't fit the design for them.

I've chosen to make this a boolean, following the style of the
'hide_opendocument_metadata' parameter, rather than inventing
some sort of 'admin context' which is open to misinterpretation.