Add Youtube Embed Analytics
We need to be able to track when users are viewing the livestream on GOV.UK.

We should use the YouTube API to send data to GA.

https://developers.google.com/youtube/iframe_api_reference#Retrieving_playlist_information

It will help us determine if it is a useful feature to include on GOV.UK.

Most commonly people use GTM to track videos on sites but we have security concerns about this.

Using the Youtube API we should be able to get around this but it is more complicated. It will give us more accurate information about viewing numbers and how users interact with the video, E.g do users pause it, start watching late and watch at increased speed? Currently we are unable to do so and at best we can estimate the number of users but it is not accurate. For us to work this out we'd need this data to be passed to GA or equivalent as events.

It's done when we can accurately record how many users viewed the live stream at any point.

It would be nice to be able to record how users are interacting with the video with events to trigger during the course of the stream.

https://trello.com/c/zAQSVUqi/124-add-livestream-tracking