Update hint on radio buttons
Using a skip to content link on a page with a radio button with page heading
and hint causes VoiceOver on Mac to read out the hint text out automatically.
I think that this is caused by the `aria-describedby` attribute - but pointing
that attribute at the heading caused nothing to be read out when the skip to
content link was used.

Using `aria-describedby` seems to have made the experience worse. Bearing in
mind the [First rule of ARIA use](https://www.w3.org/TR/using-aria/#firstrule)
I've removed the attribute. This means that nothing is automatically read out
in VoiceOver, which I think is preferable to a hint with no context.