Load application styles after gem styles
In frontend applications, styles/scripts from the gem are loaded first, then the application-level styles/scripts which sometimes depend or inherit from the ones in the gem. With this change we apply the same order to the component guide so that components look and behave the same as in app.