Adapt phase banner to allow embedding govspeak
This changes the phase banner to allow a govspeak component to be
embedded and still render ok.

To do this I've made changes to the HTML structure of the phase
component so that it no longer is only inline elements in a p element.
This allows the govspeak component to be inserted without the HTML
becoming invalid.

This component only appears correct with govspeak when the small and
unspaced modifiers are applied.