Build with sassc
This removes the sass depedency of this gem and instead replaces it with
a sassc-rails dependency. This is set as a development dependency as
this is only required for the dummy version of this app and this gem
should not force a particular version of sass on consuming apps without
good reason.

This should help GOV.UK gradually move towards a world without Ruby
SASS. Prior to this change introducing govuk_publishing_components into
an application forced Ruby SASS in as a dependency, this has the effect
in Sprockets of choosing Ruby SASS over sassc.

A downside of doing this is that it will make the warm cache run slower
for requests. Until GOV.UK Frontend resolves some performance issues [1]
we may find each page of the dummy app loads a reasonable amount slower
in a development environment. Although not ideal, I don't think this
performance issue will cause anyone significant problems developing on
this app (I get about 2.5s on my local machine) and it'll be nice for us
to have already started the sassc journey.

[1]: https://github.com/alphagov/govuk-frontend/issues/1671