Enable building CSS via sassc
This disables the css_compressor as this does not work with
govuk-frontend. To go into details, the sprockets SASS compression
process is slightly strange in that it converts the .scss file into
uncompressed CSS in one run and then in a second run it uses SASS to
compress it. This process causes a problem if we have any CSS functions
that conflict with SASS functions - as is the case with
[GOV.UK Frontend _width-container][]. We hit a problem with this where the
unquote function is removed on the first SASS run and causes a conflict
on the second.

To make up for the disabled compressor we set SASS to output as
compressed without comments.

We also need to require "sassc-rails" in the application file as it
doesn't seem to be required automatically, requiring this removes the
need to require sprockets/railtie.

[GOV.UK Frontend _width-container]: https://github.com/alphagov/govuk-frontend/blame/cc4e3cc711b6557505d89ec46278f3abbd8c2108/src/govuk/objects/_width-container.scss#L32